Move SearchCursor.py to index folder
