+++ mcrit/index/SearchCursor.py
