+++ mcrit/storage/MongoDbStorage.py
@@ -16,13 +16,13 @@ except:
 
 from picblocks.blockhasher import BlockHasher
 
+from mcrit.index.SearchCursor import FullSearchCursor
 from mcrit.index.SearchQueryParser import SearchConditionNode, SearchTermNode
 from mcrit.libs.utility import generate_unique_groups
 from mcrit.storage.FamilyEntry import FamilyEntry
 from mcrit.storage.FunctionEntry import FunctionEntry
 from mcrit.storage.MatchingCache import MatchingCache
 from mcrit.storage.SampleEntry import SampleEntry
-from mcrit.storage.SearchCursor import FullSearchCursor
 from mcrit.storage.StorageInterface import StorageInterface
 
 if TYPE_CHECKING: # pragma: no cover
